qr_inventory/requirements.txt
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.2.0
crispy-bootstrap4==2024.1
cssselect2==0.8.0
Django==5.0.3
django-cors-headers==4.7.0
django-crispy-forms==2.1
djangorestframework==3.15.0
docx==0.2.4
Flask==3.1.1
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
pdfkit==1.0.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-docx==1.1.2
pytz==2024.1
qrcode==8.2
reportlab==4.4.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.6.0
zopfli==0.2.3.post1
